// frontend/src/client/types.gen.ts
export type TaskCreate = {
    title: string;
    description?: (string | null);
    priority?: TaskPriority;
    status?: TaskStatus;
    due_date?: (string | null);
};

export type TaskPriority = 'low' | 'medium' | 'high' | 'urgent';

export type TaskPublic = {
    title: string;
    description?: (string | null);
    priority?: TaskPriority;
    status?: TaskStatus;
    due_date?: (string | null);
    id: string;
    created_at: string;
    updated_at: string;
    owner_id: string;
};

export type TasksPublic = {
    data: Array<TaskPublic>;
    count: number;
};

export type TaskStatus = 'todo' | 'in_progress' | 'done' | 'cancelled';